requirements.txt
Django~=5.1
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1